requirements.txt
langchain 
langchain-core
langchain-openai 
langchain-google-genai 
pydantic
google-cloud-storage
google-cloud-bigquery
python-dotenv
pyyaml
tiktoken